Add jepsen tests and clojure client for jepsen
This folder contains Jepsen tests for Xenon and Xenon Clojure client to help
write Jepsen tests.

Jepsen is a framework for distributed systems verification, with fault
injection, written by Kyle Kingsbury.  It fuzzes the system with random
operations while injecting network partitions.  The results of operation
history is analyzed to see if the system violates any of the consistency
properties it claims to have.  It generates graphs of performance and
availability, helping user characterize how a system responds to different
faults.

We ran a Jepsen test, as part of our Xenon testing, to determine if any
consistency issues could be uncovered.  In our testing of Read, Write and CAS
(Compare-And-Set) operations, we found them to be linearizable, and Xenon
gracefully recovered from partitions without introducing any consistency
issues.

Change-Id: Iae06c9197f750633ed70e4274705e0bd238058c6