Add richer swagger support.
   https://www.pivotaltracker.com/story/show/145124239

This is in the context of a change which was pushed to xenon-common to add
richer ServiceDocumentDescriptionSupport, on which this relies.

As such have modified dependency to be on xenon-1.5.1-SNAPSHOT and
added its repository - this is to be modified to the official release
when made.

Includes better support for:

- specific HTTP verbs
   - responses, produces, consumes support

- externalized HTML descriptions

- example values for types other than String

- filtering which service endpoints to expose based upon their
  declared support level (PRIVATE,DEPRECATED,PUBLIC)

Have also removed service-specific implementation of Accept-Encoding: gzip
as this has been added to xenon core.

Change-Id: Iee17ad73a74471545aac64e60c408e2befa567a0