Correcting swagger type for generic maps
All map types were being exposed as a map of
strings to strings in the swagger spec. This
adds a special case for Map<String, Object>,
as well as correctly generating the map type
for other types.

Also reverting some changes in the GET
handler for the descriptor service because
/? may not work with all applications.

Change-Id: I97b2491f32b1fca8d6330e0aef7865496addd53e