Overhaul distributed-tracing to be Opentracing based.
* Overhaul and update to OpenTracing rather than Brave v2 libraries.
This is a major API incompatible break but should +  be unlikely to
happen in future, as OpenTracing has passed 1.0 in its API and is very
much looking for long term stability to enhance adoption.

* Added support for Zipkin V2 API servers in addition to V1 servers.

* Added support for Jaeger reporting.

* Rewrite README to cover the new features and be more direct.

* Support tracing sendRequest and handleRequest implicitly via a hacked
  up ServiceHost subclass.

This requires ServiceHost to make a couple of methods/attributes more
visible to permit sensible overriding, thus a xenon dependency version
bump.

Current status is a working demo of pervasive tracing through stateless
and stateful services, but requires extensive subclassing of ServiceHost
: useful as a demonstration, but not ready for consumption.

Pushing this up for folk to comment and discuss, while I start on deeper
integration to address these issues.

Change-Id: Ide556c24c4df17d8e78c19094cfd0a6df835d0bf