Minimal reproduction of issue with NestCompletion in Host.sendRequest
I was hoping to override sendRequest to re-establish opentracing
contexts around the handling code, but my should-be-correct code results
in service host startup failures with timeouts.

Run PropogationTest which should (with the code in the repo) error
due to a 404 on /stateless, but instead stalls at service host startup
and then timesout.

Change-Id: I844486df3cdbc93f9340ecf29b34a54b311aa7b0