Repro: SupportLevel.INTERNAL isn't hidden when PUBLIC level set
I'm trying to flush out swagger docs for my service and came across a
bug. I added a "supportLevel = SupportLevel.INTERNAL" to one of my
APIs... and then ensured the SwaggerDescriptorService used PUBLIC for
its level.

I expected this INTERNAL service to not show up in the generated Swagger
docs. However, it does show up.

If I set my Route annotation to NOT_SUPPORTED, then I see the API is
removed from the Swagger docs... but I think marking the API as INTERNAL
is much more appropriate for my use case vs NOT_SUPPORTED.

The issue is that when Xenon generates the document template, it has
access to the Service class instance... so it can inspect the
RouteDocumentation annotation. Xenon, by defaults, hides all
NOT_SUPPORTED docs from the generated template via the
ServiceDocumentDescriptionHelper class.

However, the SwaggerDescriptionService (and its associated
SwaggerAssembler) doesn't have access to the Service class (nor its
associated RouteDocumentation annotations. Xenon doesn't populate the
SupportLevel in the generated template either... so when the
SwaggerAssembler looks at the supportLevel (on line 735)... it's always
going to be null... which is why it can't hide anything.

Change-Id: I26862afc577fa36a358342fcad0166a53fec58e1